+++ python/pyautd3/__init__.py
@@ -55,4 +55,4 @@ __all__ = [
     "TimerStrategy",
 ]
 
-__version__ = "13.0.0"
+__version__ = "14.0.1"