+++ python/pyautd3/stm/stm.py
@@ -15,6 +15,7 @@ from abc import ABCMeta, abstractmethod
 import functools
 from typing import Optional, List, Tuple, Union
 from collections.abc import Iterable
+import ctypes
 
 import numpy as np
 
@@ -103,17 +104,19 @@ class FocusSTM(STM):
         self._points = []
         self._duty_shifts = []
 
-    def ptr(self, geometry: Geometry) -> DatagramBodyPtr:
-        points = np.ctypeslib.as_ctypes(np.array(self._points).astype(np.double))
-        shifts = np.ctypeslib.as_ctypes(np.array(self._duty_shifts).astype(np.uint8))
+    def ptr(self, _: Geometry) -> DatagramBodyPtr:
+        points = np.ctypeslib.as_ctypes(np.array(self._points).astype(ctypes.c_double))
+        shifts = np.ctypeslib.as_ctypes(
+            np.array(self._duty_shifts).astype(ctypes.c_uint8)
+        )
         return Base().focus_stm(self.props(), points, shifts, len(self._duty_shifts))
 
     @staticmethod
-    def from_sampling_frequency(sampling_freq: float) -> "FocusSTM":
+    def with_sampling_frequency(sampling_freq: float) -> "FocusSTM":
         return FocusSTM(None, sampling_freq, None)
 
     @staticmethod
-    def from_sampling_frequency_division(sampling_freq_div: int) -> "FocusSTM":
+    def with_sampling_frequency_division(sampling_freq_div: int) -> "FocusSTM":
         return FocusSTM(None, None, sampling_freq_div)
 
     def add_focus(self, point: np.ndarray, duty_shift: int = 0) -> "FocusSTM":