+++ src/autd3-protobuf/src/traits.rs
@@ -4,7 +4,7 @@
  * Created Date: 30/06/2023
  * Author: Shun Suzuki
  * -----
- * Last Modified: 10/07/2023
+ * Last Modified: 12/07/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2023 Shun Suzuki. All rights reserved.
@@ -230,6 +230,97 @@ impl ToMessage for autd3::gain::TransducerTest {
                         .collect(),
                 })),
             })),
+        }
+    }
+}
+
+impl<T: autd3_core::geometry::Transducer> ToMessage for autd3::gain::Grouped<T> {
+    type Message = Grouped;
+
+    fn to_msg(&self) -> Self::Message {
+        Self::Message {
+            groups: self
+                .gain_map()
+                .keys()
+                .map(|&k| {
+                    let gain = if let Some(g) = self.get_gain::<autd3::gain::Focus>(k) {
+                        match g.to_msg().datagram.unwrap() {
+                            datagram::Datagram::Gain(g) => Some(g),
+                            _ => None,
+                        }
+                    } else if let Some(g) = self.get_gain::<autd3::gain::Bessel>(k) {
+                        match g.to_msg().datagram.unwrap() {
+                            datagram::Datagram::Gain(g) => Some(g),
+                            _ => None,
+                        }
+                    } else if let Some(g) = self.get_gain::<autd3::gain::Null>(k) {
+                        match g.to_msg().datagram.unwrap() {
+                            datagram::Datagram::Gain(g) => Some(g),
+                            _ => None,
+                        }
+                    } else if let Some(g) = self.get_gain::<autd3::gain::Plane>(k) {
+                        match g.to_msg().datagram.unwrap() {
+                            datagram::Datagram::Gain(g) => Some(g),
+                            _ => None,
+                        }
+                    } else if let Some(g) = self.get_gain::<autd3::gain::TransducerTest>(k) {
+                        match g.to_msg().datagram.unwrap() {
+                            datagram::Datagram::Gain(g) => Some(g),
+                            _ => None,
+                        }
+                    } else if let Some(g) =
+                        self.get_gain::<autd3_gain_holo::SDP<autd3_gain_holo::NalgebraBackend>>(k)
+                    {
+                        match g.to_msg().datagram.unwrap() {
+                            datagram::Datagram::Gain(g) => Some(g),
+                            _ => None,
+                        }
+                    } else if let Some(g) =
+                        self.get_gain::<autd3_gain_holo::EVP<autd3_gain_holo::NalgebraBackend>>(k)
+                    {
+                        match g.to_msg().datagram.unwrap() {
+                            datagram::Datagram::Gain(g) => Some(g),
+                            _ => None,
+                        }
+                    } else if let Some(g) =
+                        self.get_gain::<autd3_gain_holo::Naive<autd3_gain_holo::NalgebraBackend>>(k)
+                    {
+                        match g.to_msg().datagram.unwrap() {
+                            datagram::Datagram::Gain(g) => Some(g),
+                            _ => None,
+                        }
+                    } else if let Some(g) =
+                        self.get_gain::<autd3_gain_holo::GS<autd3_gain_holo::NalgebraBackend>>(k)
+                    {
+                        match g.to_msg().datagram.unwrap() {
+                            datagram::Datagram::Gain(g) => Some(g),
+                            _ => None,
+                        }
+                    } else if let Some(g) =
+                        self.get_gain::<autd3_gain_holo::GSPAT<autd3_gain_holo::NalgebraBackend>>(k)
+                    {
+                        match g.to_msg().datagram.unwrap() {
+                            datagram::Datagram::Gain(g) => Some(g),
+                            _ => None,
+                        }
+                    } else if let Some(g) =
+                        self.get_gain::<autd3_gain_holo::LM<autd3_gain_holo::NalgebraBackend>>(k)
+                    {
+                        match g.to_msg().datagram.unwrap() {
+                            datagram::Datagram::Gain(g) => Some(g),
+                            _ => None,
+                        }
+                    } else if let Some(g) = self.get_gain::<autd3_gain_holo::Greedy>(k) {
+                        match g.to_msg().datagram.unwrap() {
+                            datagram::Datagram::Gain(g) => Some(g),
+                            _ => None,
+                        }
+                    } else {
+                        None
+                    };
+                    grouped::Group { id: k as _, gain }
+                })
+                .collect(),
         }
     }
 }
@@ -594,7 +685,7 @@ impl FromMessage<TransducerTest> for autd3::gain::TransducerTest {
     }
 }
 
-impl<'a, T: autd3_core::geometry::Transducer> FromMessage<Grouped> for autd3::gain::Grouped<'a, T> {
+impl<T: autd3_core::geometry::Transducer> FromMessage<Grouped> for autd3::gain::Grouped<T> {
     #[allow(clippy::unnecessary_cast)]
     fn from_msg(msg: &Grouped) -> Self {
         msg.groups.iter().fold(Self::new(), |acc, v| {
@@ -897,7 +988,7 @@ impl FromMessage<FocusStm> for autd3_core::stm::FocusSTM {
     }
 }
 
-impl<'a, T: autd3_core::geometry::Transducer + 'a> FromMessage<GainStm>
+impl<'a, T: autd3_core::geometry::Transducer + 'a + 'static> FromMessage<GainStm>
     for autd3_core::stm::GainSTM<'a, T>
 {
     fn from_msg(msg: &GainStm) -> Self {
@@ -915,10 +1006,8 @@ impl<'a, T: autd3_core::geometry::Transducer + 'a> FromMessage<GainStm>
                     Box::new(autd3::prelude::TransducerTest::from_msg(msg))
                         as Box<dyn autd3_core::gain::Gain<T>>
                 }
-                gain::Gain::Grouped(msg) => {
-                    Box::new(autd3::prelude::Grouped::<'a, T>::from_msg(msg))
-                        as Box<dyn autd3_core::gain::Gain<T>>
-                }
+                gain::Gain::Grouped(msg) => Box::new(autd3::prelude::Grouped::<T>::from_msg(msg))
+                    as Box<dyn autd3_core::gain::Gain<T>>,
                 gain::Gain::Sdp(msg) => Box::new(autd3_gain_holo::SDP::from_msg(msg))
                     as Box<dyn autd3_core::gain::Gain<T>>,
                 gain::Gain::Evp(msg) => Box::new(autd3_gain_holo::EVP::from_msg(msg))